v2:rpi:canvas2d:Use boundarybox for clipping space
This boundarybox checks if the CanvasDrawable2d states it is inside the
clipping space. If the CanvasDrawable2d is not inside the clipping space
we never have to consider it again. Note that the CanvasDrawable2d is
not guaranteed to return false if it is outside of clipping space. But
this is better than nothing at all